Add files via upload
The ipynb is a file of python code I created to examine the data we were provided in excel format (Healthcare_dataset.xlsx)  - it creates a pandas dataframe from the data page of this excel file --it also adds a more useful age-related column 'Age_Category' (without removing the 'Age_Bucket column) -- then it converts the pandas datafile back to a csv which gets saved. The other uploaded file is the resulting csv. I did not find any missing data. There are many categorical variables (mostly flags in the form of y/n responses).