Add files via upload
Hi Team,
I created bar charts between the independent variables and the target variable Drug_Persistency to get a preliminary impression of possible correlations. Here is my Python file I created on Jupyter, as well as a PDF that shows the code and the output.  Note that one column name needed to be corrected (to eliminate a comma that created trouble). The bar charts make it easy to see that many variables will be able to be eliminated before running the ML classification programs. We will also have to make some determinations about which variables may not be independent from each other (some might need to be eliminated for that reason - as a prereq for some of the classification algorithms).  I also think we need to create more useful buckets for some of the variables.  Feel free to play with or amend my code as you see fit.
Jennifer