Add files via upload
Note from Jennifer:
Hi, I did some work on our created .csv file (the one on our main branch that is a version of our health data file originally in a slightly different format). 

Here is a description of what is in this .ipynb file I am uploading now:
*Two more corrections for small errors in the data file: one column name had a comma in the name (which was interfering with creation of bar charts etc.) and in the Ntm_Speciality column there was an OB &GYN entry improperly entered, which is now included correctly with the others that are 'obstetrics & gynecology.'

*I also created a new column with different Ntm Speciality buckets - I think these buckets are superior to the ones in the original data file. 

*Also, I narrowed down the list of desirable features to use in the machine learning classification process, based on the bar charts, and created a list of column names we are using,  calling it features (you will find it defined in the .ipynb file).  Note that it is not appropriate to have many dependent columns when building these models (it contradicts requirements for building the model) - so for example, the only Ntm speciality column being used is now my new MyBuckets version. I'm confident this list of the predictor/independent variables will get smaller as we start building and can refine the model. 

The next step is converting these feature columns - the y/n should be coded instead as 1/0 and some of the other columns need to be re-coded or written up using dummy variables then changed to 1/0 for those dummy variables. 

I'm sharing this now so you can look at it while it is at this stage.

If someone wants to write the code to convert y/n to 1/0 or create dummy variables, let me know (if I don't hear from anyone I will do that tomorrow).

Feel free to edit any of this file - or if you see problems, let me know.
Thanks.